\name{myarray}
\alias{myarray}
\title{The myarray class}
\usage{
myarray(data, dims)
}
\arguments{
\item{data}{A vector of data elements.}
\item{dims}{A vector of array dimensions.}
}
\value{
An object of class \code{myarray}.
}
\description{
Create an object of class \code{myarray}.
}
\author{
Jon Clayden
}
